projects/dynamic-form/src/lib/components/radio/radio.component.ts
import { Component, Input } from '@angular/core';
import { FormControl } from '@angular/forms';
import { RadioConfiguration } from '../../models/models';
import { errorControlMessage } from '../../utils/errors/error-control';
import { hasRequiredValidator } from '../../utils/validators/has-validator';

@Component({
  selector: 'talan-radio',
  templateUrl: './radio.component.html',
  styleUrls: ['./radio.component.scss'],
})
export class RadioComponent {
  private _field!: RadioConfiguration;

  get field(): RadioConfiguration {
    return this._field;
  }

  @Input() set field(value: RadioConfiguration) {
    this._field = value;
  }

  @Input() control!: FormControl;

  get required() {
    return hasRequiredValidator(this.control);
  }

  constructor() {}

  errorControlMessage = errorControlMessage;
}
